refactor out base tls mirror connection
