+++ transport/internet/tlsmirror/mirrorbase/base.go
@@ -1,0 +1,3 @@
+package mirrorbase
+
+//go:generate go run github.com/v2fly/v2ray-core/v5/common/errors/errorgen
