+++ transport/internet/tlsmirror/mirrorbase/errors.generated.go
@@ -1,0 +1,9 @@
+package mirrorbase
+
+import "github.com/v2fly/v2ray-core/v5/common/errors"
+
+type errPathObjHolder struct{}
+
+func newError(values ...interface{}) *errors.Error {
+	return errors.New(values...).WithPathObj(errPathObjHolder{})
+}
